refactor: make Table component reusable and integrate it into Reports screen
